+++ src/Ratchet/Session/SessionProvider.php
@@ -1,8 +1,8 @@
 <?php
 namespace Ratchet\Session;
-use Ratchet\MessageComponentInterface;
 use Ratchet\ConnectionInterface;
-use Ratchet\WebSocket\WsServerInterface;
+use Ratchet\Http\HttpServerInterface;
+use Psr\Http\Message\RequestInterface;
 use Ratchet\Session\Storage\VirtualSessionStorage;
 use Ratchet\Session\Serialize\HandlerInterface;
 use Symfony\Component\HttpFoundation\Session\Session;
@@ -14,7 +14,7 @@ use Symfony\Component\HttpFoundation\Session\Storage\Handler\NullSessionHandler;
  * Your website must also use Symfony HttpFoundation Sessions to read your sites session data
  * If your are not using at least PHP 5.4 you must include a SessionHandlerInterface stub (is included in Symfony HttpFoundation, loaded w/ composer)
  */
-class SessionProvider implements MessageComponentInterface, WsServerInterface {
+class SessionProvider implements HttpServerInterface {
     /**
      * @var \Ratchet\MessageComponentInterface
      */
@@ -38,13 +38,13 @@ class SessionProvider implements MessageComponentInterface, WsServerInterface {
     protected $_serializer;
 
     /**
-     * @param \Ratchet\MessageComponentInterface          $app
+     * @param \Ratchet\Http\HttpServerInterface           $app
      * @param \SessionHandlerInterface                    $handler
      * @param array                                       $options
      * @param \Ratchet\Session\Serialize\HandlerInterface $serializer
      * @throws \RuntimeException
      */
-    public function __construct(MessageComponentInterface $app, \SessionHandlerInterface $handler, array $options = array(), HandlerInterface $serializer = null) {
+    public function __construct(HttpServerInterface $app, \SessionHandlerInterface $handler, array $options = array(), HandlerInterface $serializer = null) {
         $this->_app     = $app;
         $this->_handler = $handler;
         $this->_null    = new NullSessionHandler;
@@ -70,8 +70,20 @@ class SessionProvider implements MessageComponentInterface, WsServerInterface {
     /**
      * {@inheritdoc}
      */
-    function onOpen(ConnectionInterface $conn) {
-        if (!isset($conn->WebSocket) || null === ($id = $conn->WebSocket->request->getCookie(ini_get('session.name')))) {
+    public function onOpen(ConnectionInterface $conn, RequestInterface $request = null) {
+        $sessionName = ini_get('session.name');
+
+        $id = array_reduce($request->getHeader('Cookie'), function($accumulator, $cookie) use ($sessionName) {
+            if ($accumulator) {
+                return $accumulator;
+            }
+
+            $crumbs = $this->parseCookie($cookie);
+
+            return isset($crumbs['cookies'][$sessionName]) ? $crumbs['cookies'][$sessionName] : false;
+        }, false);
+
+        if (null === $request || false === $id) {
             $saveHandler = $this->_null;
             $id = '';
         } else {
@@ -84,7 +96,7 @@ class SessionProvider implements MessageComponentInterface, WsServerInterface {
             $conn->Session->start();
         }
 
-        return $this->_app->onOpen($conn);
+        return $this->_app->onOpen($conn, $request);
     }
 
     /**
@@ -108,17 +120,6 @@ class SessionProvider implements MessageComponentInterface, WsServerInterface {
      */
     function onError(ConnectionInterface $conn, \Exception $e) {
         return $this->_app->onError($conn, $e);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getSubProtocols() {
-        if ($this->_app instanceof WsServerInterface) {
-            return $this->_app->getSubProtocols();
-        } else {
-            return array();
-        }
     }
 
     /**
@@ -157,5 +158,86 @@ class SessionProvider implements MessageComponentInterface, WsServerInterface {
      */
     protected function toClassCase($langDef) {
         return str_replace(' ', '', ucwords(str_replace('_', ' ', $langDef)));
+    }
+
+    /**
+     * Taken from Guzzle3
+     */
+    private static $cookieParts = array(
+        'domain'      => 'Domain',
+        'path'        => 'Path',
+        'max_age'     => 'Max-Age',
+        'expires'     => 'Expires',
+        'version'     => 'Version',
+        'secure'      => 'Secure',
+        'port'        => 'Port',
+        'discard'     => 'Discard',
+        'comment'     => 'Comment',
+        'comment_url' => 'Comment-Url',
+        'http_only'   => 'HttpOnly'
+    );
+
+    /**
+     * Taken from Guzzle3
+     */
+    private function parseCookie($cookie, $host = null, $path = null, $decode = false) {
+        // Explode the cookie string using a series of semicolons
+        $pieces = array_filter(array_map('trim', explode(';', $cookie)));
+
+        // The name of the cookie (first kvp) must include an equal sign.
+        if (empty($pieces) || !strpos($pieces[0], '=')) {
+            return false;
+        }
+
+        // Create the default return array
+        $data = array_merge(array_fill_keys(array_keys(self::$cookieParts), null), array(
+            'cookies'   => array(),
+            'data'      => array(),
+            'path'      => $path ?: '/',
+            'http_only' => false,
+            'discard'   => false,
+            'domain'    => $host
+        ));
+        $foundNonCookies = 0;
+
+        // Add the cookie pieces into the parsed data array
+        foreach ($pieces as $part) {
+
+            $cookieParts = explode('=', $part, 2);
+            $key = trim($cookieParts[0]);
+
+            if (count($cookieParts) == 1) {
+                // Can be a single value (e.g. secure, httpOnly)
+                $value = true;
+            } else {
+                // Be sure to strip wrapping quotes
+                $value = trim($cookieParts[1], " \n\r\t\0\x0B\"");
+                if ($decode) {
+                    $value = urldecode($value);
+                }
+            }
+
+            // Only check for non-cookies when cookies have been found
+            if (!empty($data['cookies'])) {
+                foreach (self::$cookieParts as $mapValue => $search) {
+                    if (!strcasecmp($search, $key)) {
+                        $data[$mapValue] = $mapValue == 'port' ? array_map('trim', explode(',', $value)) : $value;
+                        $foundNonCookies++;
+                        continue 2;
+                    }
+                }
+            }
+
+            // If cookies have not yet been retrieved, or this value was not found in the pieces array, treat it as a
+            // cookie. IF non-cookies have been parsed, then this isn't a cookie, it's cookie data. Cookies then data.
+            $data[$foundNonCookies ? 'data' : 'cookies'][$key] = $value;
+        }
+
+        // Calculate the expires date
+        if (!$data['expires'] && $data['max_age']) {
+            $data['expires'] = time() + (int) $data['max_age'];
+        }
+
+        return $data;
     }
 }