+++ src/Ratchet/WebSocket/MessageCallableInterface.php
@@ -1,0 +1,8 @@
+<?php
+namespace Ratchet\WebSocket;
+use Ratchet\ConnectionInterface;
+use Ratchet\RFC6455\Messaging\MessageInterface;
+
+interface MessageCallableInterface {
+    public function onMessage(ConnectionInterface $conn, MessageInterface $msg);
+}