+++ src/Ratchet/WebSocket/MessageComponentInterface.php
@@ -1,0 +1,6 @@
+<?php
+namespace Ratchet\WebSocket;
+use Ratchet\ComponentInterface;
+
+interface MessageComponentInterface extends ComponentInterface, MessageCallableInterface {
+}